encapsulating producer
