+++ client/Producer.jsx
@@ -1,25 +1,13 @@
 import React, { useState, useEffect } from "react";
 import { render } from 'react-dom'
 import io from 'socket.io-client'
-// const io = require('socket.io-client')
+
 
 const socket = io("http://localhost:3001");
 
 function Producer() {
   const [num, setNum] = useState(30);
-  
-  // useEffect(() => {
-  //   // console.log('In useeffect of producer component')
-  //   // socket.emit('postMessage', {
-  //   //   key: String(3),
-  //   //   value: String(num)
-  //   // })
-  
-  //   // return () => {
-  //   //   console.log("is Producer ever off?");
-  //   //   socket.off();
-  //   // }
-  // });
+
 
   function socketProducerInvoke() {
     setNum(num + 1);
